[okteta-minidump] Mirror arm cpsr modes with bit 4
In the Current Program Status Register the mode was previously five bits
and the high bit was always set. In some versions of aarch the mode is
documented to be four bits with the high bit now reserved. Until this
bit is re-assigned, interpret this high bit as an ignorable part of the
mode.