[okteta-minidump] Treat CrashpadInfo as variable length
The CrashpadInfo looks like a standard fixed size struct but additional
fields have been added to it in the past and may be added in the future.
Add known fields until reaching the declared size of the stream.

Before this change Okteta would just keep reading data and report
arbitrary following data for any fields which weren't actually in the
stream.