Add Kaitai Minidump structure
The Minidump format is used for serializing the (partial) state of a
process, often for crash reporting. The format is by defined Microsoft
for use with Windows and other Microsoft software. However, it is now
more generally used due to the Breakpad and Crashpad projects, which
have extended the format so that it can be used with more platforms and
operating systems.

This inital commit contains support for the most common parts of the
format, including thread context parsers for arm, arm64, x86, and amd64.