[kaitai-minidump] Add labels for webide
The Kaitai WebIDE [0] allows for structures to be annotated with
`-webide-representation` to provide a formatted label. This is
particularly useful with arrays of structures where each structure can
be annotated with a relevant key so the user is not required to click on
a large number of twiddles just to get an idea of what is in the array.
This is used in conjunction with `-webide-parse-mode: eager` to ensure
that the key values are read before the parent structure attempts to use
the key values in its formatted label.

The `-webide-representation` value is a yaml string parsed like

    webide-representation: (.* "{" substitution "}")* .*
    substitution: ident (":" formatter)?
    ident: id ("." id)?
    id: "this"|"_parent"|<member>
    formatter: format ("," format)?
    format: ("sep"|"str"|"hex"|"dec"|"uuid"|"flags")("=" .*)?

The uuid format by default reads the values big-endian. To format
correctly it must be specified as uuid=ms. Unfortunately, the uuid
format only works on raw bytes (cannot be applied on top of structured
data) so the minidump_guid structure is simplified.

[0] https://ide.kaitai.io/devel/